app/requirements.txt
beautifulsoup4==4.12.3
customtkinter==5.2.2
mysql_connector_repackaged==0.3.1
pandas==2.2.2
Pillow==10.3.0
Requests==2.32.3
